backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
groq
google-generativeai
requests
openai-whisper
markdown-it-py
pydantic[email]
anthropic
deepgram-sdk